+++ README.md
@@ -1,6 +1,39 @@
+[![Starware](https://img.shields.io/badge/Starware-⭐-black?labelColor=f9b00d)](https://github.com/zepfietje/starware)
+<!-- ALL-CONTRIBUTORS-BADGE:START - Do not remove or modify this section -->
+[![All Contributors](https://img.shields.io/badge/all_contributors-11-orange.svg?style=flat-square)](#contributors-)
+<!-- ALL-CONTRIBUTORS-BADGE:END -->
 
 # About App
 - The project repository of the mobile app development for Institute Counselling Services, NIT Rourkela.
+
+
+## Contributing 🎃
+------------
+
+This repository is one of the many repositories maintained by the NIT Rourkela tech community. To get more info about all of our projects, visit our [Info Doc](https://www.notion.so/c019f8d965c24047b92f227a1b20fe4b?v=b1de077e3ea54a7daf480e8ca59e3167) <br>
+We are super happy that you are here and invite you to join us at [our Hacktoberfest Meetup](http://bit.ly/NITR-HF). <br>
+Our Slack Community: [Slack Invite](http://bit.ly/NITRDevs) <br>
+#### We are are excited to see your awesome PRs. As a recognition of your awesome efforts, we would be giving a badge of gratitute to you to showcase your fantastic contribution on sending your first PR in October 2020 
+`Contributions for Hacktoberfest 2020 are welcome 🎉🎉`
+
+Please refer to the project's style and contribution guidelines for submitting patches and additions. In general, we follow the "fork-and-pull" Git workflow.
+
+ 1. **Fork** the repo on GitHub
+ 2. **Clone** the project to your own machine
+ 3. **Commit** changes to your own branch
+ 4. **Push** your work back up to your fork
+ 5. Submit a **Pull request** so that we can review your changes
+
+NOTE 1: Please abide by the [Contributing Guidelines](https://github.com/dscnitrourkela/project-avocado/blob/master/CONTRIBUTING.md).
+
+NOTE 2: Please abide by the [Code of Conduct](https://github.com/dscnitrourkela/project-avocado/blob/master/CODE_OF_CONDUCT.md).
+
+
+## Starware
+
+dscnitrourkela/project-avocado is Starware.  
+This means you're free to use the project, as long as you star its GitHub repository.  
+Your appreciation makes us grow and glow up. ⭐
 
 ## Preview
 <pre>
@@ -12,3 +45,34 @@
 - Timetable selector
 - Manage your mentors
 - Get Notifications
+
+## Contributors ✨
+
+Thanks goes to these wonderful people ([emoji key](https://allcontributors.org/docs/en/emoji-key)):
+
+<!-- ALL-CONTRIBUTORS-LIST:START - Do not remove or modify this section -->
+<!-- prettier-ignore-start -->
+<!-- markdownlint-disable -->
+<table>
+  <tr>
+    <td align="center"><a href="https://designrknight-website.web.app/"><img src="https://avatars0.githubusercontent.com/u/27865704?v=4" width="100px;" alt=""/><br /><sub><b>Abel Mathew</b></sub></a><br /><a href="https://github.com/dscnitrourkela/project-avocado/commits?author=DesignrKnight" title="Code">💻</a> <a href="#projectManagement-DesignrKnight" title="Project Management">📆</a> <a href="#maintenance-DesignrKnight" title="Maintenance">🚧</a></td>
+    <td align="center"><a href="https://allcontributors.org"><img src="https://avatars1.githubusercontent.com/u/46410174?v=4" width="100px;" alt=""/><br /><sub><b>All Contributors</b></sub></a><br /><a href="#tool-all-contributors" title="Tools">🔧</a></td>
+    <td align="center"><a href="https://www.linkedin.com/in/patibanda-sriram-237a0b193/"><img src="https://avatars3.githubusercontent.com/u/56649197?v=4" width="100px;" alt=""/><br /><sub><b>SriramPatibanda</b></sub></a><br /><a href="https://github.com/dscnitrourkela/project-avocado/commits?author=SriramPatibanda" title="Code">💻</a></td>
+    <td align="center"><a href="https://github.com/nayakastha"><img src="https://avatars0.githubusercontent.com/u/58568514?v=4" width="100px;" alt=""/><br /><sub><b>Astha Nayak</b></sub></a><br /><a href="https://github.com/dscnitrourkela/project-avocado/commits?author=nayakastha" title="Code">💻</a></td>
+    <td align="center"><a href="https://github.com/Thesmader"><img src="https://avatars2.githubusercontent.com/u/34758667?v=4" width="100px;" alt=""/><br /><sub><b>Smarak Das</b></sub></a><br /><a href="https://github.com/dscnitrourkela/project-avocado/commits?author=Thesmader" title="Code">💻</a> <a href="#projectManagement-Thesmader" title="Project Management">📆</a> <a href="#maintenance-Thesmader" title="Maintenance">🚧</a></td>
+    <td align="center"><a href="https://github.com/ankank30"><img src="https://avatars3.githubusercontent.com/u/35187467?v=4" width="100px;" alt=""/><br /><sub><b>Ankesh Anku</b></sub></a><br /><a href="https://github.com/dscnitrourkela/project-avocado/commits?author=ankank30" title="Code">💻</a> <a href="#projectManagement-ankank30" title="Project Management">📆</a></td>
+    <td align="center"><a href="https://github.com/Chinmay-KB"><img src="https://avatars0.githubusercontent.com/u/13520364?v=4" width="100px;" alt=""/><br /><sub><b>Chinmay Kabi</b></sub></a><br /><a href="https://github.com/dscnitrourkela/project-avocado/commits?author=Chinmay-KB" title="Code">💻</a> <a href="#projectManagement-Chinmay-KB" title="Project Management">📆</a> <a href="#platform-Chinmay-KB" title="Packaging/porting to new platform">📦</a></td>
+  </tr>
+  <tr>
+    <td align="center"><a href="https://abhibhaw.team"><img src="https://avatars3.githubusercontent.com/u/39991296?v=4" width="100px;" alt=""/><br /><sub><b>Abhibhaw Asthana</b></sub></a><br /><a href="https://github.com/dscnitrourkela/project-avocado/commits?author=abhibhaw" title="Code">💻</a> <a href="#a11y-abhibhaw" title="Accessibility">️️️️♿️</a></td>
+    <td align="center"><a href="http://harshpatel.netlify.app"><img src="https://avatars1.githubusercontent.com/u/44067918?v=4" width="100px;" alt=""/><br /><sub><b>harsh patel</b></sub></a><br /><a href="https://github.com/dscnitrourkela/project-avocado/commits?author=hkp27299" title="Documentation">📖</a> <a href="#example-hkp27299" title="Examples">💡</a></td>
+    <td align="center"><a href="https://github.com/rajnis09"><img src="https://avatars2.githubusercontent.com/u/46453101?v=4" width="100px;" alt=""/><br /><sub><b>Rajnish Singh</b></sub></a><br /><a href="https://github.com/dscnitrourkela/project-avocado/commits?author=rajnis09" title="Code">💻</a></td>
+    <td align="center"><a href="https://github.com/priyanshu0405"><img src="https://avatars0.githubusercontent.com/u/58037946?v=4" width="100px;" alt=""/><br /><sub><b>priyanshu0405</b></sub></a><br /><a href="https://github.com/dscnitrourkela/project-avocado/commits?author=priyanshu0405" title="Code">💻</a></td>
+  </tr>
+</table>
+
+<!-- markdownlint-enable -->
+<!-- prettier-ignore-end -->
+<!-- ALL-CONTRIBUTORS-LIST:END -->
+
+This project follows the [all-contributors](https://github.com/all-contributors/all-contributors) specification. Contributions of any kind welcome!
