+++ scp/android/app/src/main/res/mipmap-hdpi/ic_launcher_round.png
