+++ scp/pubspec.yaml
@@ -24,6 +24,7 @@ dependencies:
   firebase_crashlytics: ^0.1.0+2
   firebase_messaging: ^5.1.6
   cloud_firestore: ^0.12.9+6
+  connectivity: ^0.4.9+3
   native_contact_dialog: ^0.2.2
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
